# openapi.yaml
openapi: 3.1.0
info:
  title: Autonomous AI Agent Safety & Permission Control System
  description: Local middleware firewall intercepting agent actions, scoring risk,
    and enforcing policy.
  version: 1.0.0
paths:
...
  /api/v1/models:
    get:
      tags:
      - Model Governance
      summary: List Models
      description: List all registered models.
      operationId: list_models_api_v1_models_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  additionalProperties: true
                  type: object
                type: array
                title: Response List Models Api V1 Models Get
      security:
      - OAuth2PasswordBearer: []
...
  /api/v1/models/{model_version}:
    get:
      tags:
      - Model Governance
      summary: Get Model Details
      description: Retrieve details of a registered model.
      operationId: get_model_details_api_v1_models__model_version__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: model_version
        in: path
        required: true
        schema:
          type: string
          title: Model Version
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Model Details Api V1 Models  Model Version  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/models/{model_version}/activate:
    post:
      tags:
      - Model Governance
      summary: Activate Model
      description: Activate a model version atomically.
      operationId: activate_model_api_v1_models__model_version__activate_post
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: model_version
        in: path
        required: true
        schema:
          type: string
          title: Model Version
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Activate Model Api V1 Models  Model Version  Activate
                  Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/models/rollback:
    post:
      tags:
      - Model Governance
      summary: Rollback Model
      description: Rollback to the most recently active/retired compatible model.
      operationId: rollback_model_api_v1_models_rollback_post
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Rollback Model Api V1 Models Rollback Post
      security:
      - OAuth2PasswordBearer: []
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /api/v1/operator/login